maktablog/requirements.txt
Django
django-bootstrap-form
Pillow
Khayyam
django-tinymce
djangorestframework
PIL
pygments
django-filter